challenge2/requirements.txt
fastapi==0.92.0
grpcio-tools==1.51.3
grpcio==1.51.3
Flask==2.2.3
